Add Custom Exception Class and embed in Controllers
Custom Eception class for index table less or equal zero

+++ Module-10/WebApplicationStudents/WebApplicationStudents/Controllers/LectorController.cs
@@ -10,6 +10,7 @@ using System.Threading.Tasks;
 using WebApplicationStudents.Validation;
 using AutoMapper;
 using WebApplicationStudents.Models;
+using WebApplicationStudents.Exceptions;
 
 namespace WebApplicationStudents.Controllers
 {
@@ -33,6 +34,8 @@ namespace WebApplicationStudents.Controllers
         [HttpGet("{id}")]
         public ActionResult<LectorBl> GetLector(int id)
         {
+            if (id <= 0) throw new IndexLessThanZeroException($"Id {id} of Lector was less than zero");
+
             return _lectorService.Get(id) switch
             {
                 null => NotFound(),
@@ -56,6 +59,8 @@ namespace WebApplicationStudents.Controllers
         [HttpPut("{id}")]
         public ActionResult<string> UpdateLector(int id, Lector lector)
         {
+            if (id <= 0) throw new IndexLessThanZeroException($"Id {id} of Lector was less than zero");
+
             var lectorBl = _mapper.Map<LectorBl>(lector) with { Id = id };
             var lectorId = _lectorService.Edit(lectorBl);
             return Ok($"api/lector/{lectorId}");
@@ -64,6 +69,8 @@ namespace WebApplicationStudents.Controllers
         [HttpDelete("{id}")]
         public ActionResult DeleteLector(int id)
         {
+            if (id <= 0) throw new IndexLessThanZeroException($"Id {id} of Lector was less than zero");
+
             _lectorService.Delete(id);
             return Ok();
         }
